refactor(core): split boot asset module
